Use size_t instead of int

+++ quicksort.h
@@ -1,7 +1,9 @@
 #ifndef QUICKSORT_H
 #define QUICKSORT_H
 
-void fquicksort(float a[], int lo, int hi, int ip[]);
-void iquicksort(int a[], int lo, int hi, int ip[]);
+#include <stddef.h>
+
+void fquicksort(float a[], size_t lo, size_t hi, size_t ip[]);
+void iquicksort(int a[], size_t lo, size_t hi, size_t ip[]);
 
 #endif /* QUICKSORT_H */
